Added a ScopedStreamPoolFactory to get rid of some duplicated code and documented all public types in the Specialized namespace.
