+++ src/SevenZip/Specialized/ArchiveReaderExtensions.cs
@@ -1,16 +1,44 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace SevenZip.Specialized;
 
+/// <summary>
+/// Extension methods for the <see cref="ArchiveReader"/> which belong
+/// to the more specific features.
+/// </summary>
 public static class ArchiveReaderExtensions
 {
     const int MaxCapacity = InMemoryArchiveRepacker.MaxCapacity;
 
+    /// <summary>
+    /// Extracts all entries that match the predicate given in the
+    /// <paramref name="extractOptions"/> and passes a bulk of
+    /// streams that were extracted in-memory to the asynchronous
+    /// <paramref name="consumeExtractResults"/> function which can
+    /// be used to read the extracted contents.
+    /// </summary>
+    /// <param name="reader">
+    /// The <see cref="ArchiveReader"/> calling this method.
+    /// </param>
+    /// <param name="consumeExtractResults">
+    /// The callback to forward the bulks of extracted files to. The
+    /// result of this asynchronous callback is awaited before the
+    /// next bulk is being extracted.
+    /// </param>
+    /// <param name="extractOptions">
+    /// Options to configure the extract operation or <c>null</c>, to
+    /// use the defaults. See <see cref="BulkExtractOptions"/> for
+    /// details.
+    /// </param>
+    /// <returns>
+    /// A task which completes once all files are extracted.
+    /// </returns>
+    /// <exception cref="InvalidOperationException">
+    /// Thrown if a file within the archive exceeds the maximum size.
+    /// </exception>
     public static async Task BulkExtract(
         this ArchiveReader reader,
         Func<IEnumerable<(ArchiveEntry, Stream)>, Task> consumeExtractResults,
@@ -19,9 +47,10 @@ public static class ArchiveReaderExtensions
         var options = extractOptions ?? new BulkExtractOptions { };
         var entries = new Dictionary<int, (ArchiveEntry, Stream)>();
 
-        using var pool = CreateScopedStreamPool(
+        using var pool = ScopedStreamPoolFactory.Create(
+            reader,
             options.UseNativeStreamPool, 
-            reader.GetStreamPoolCapacity(options.StreamPoolCapacity));
+            options.StreamPoolCapacity);
 
         using var transaction = new ExtractTransaction(reader, ArchiveFlags.None);
 
@@ -73,7 +102,7 @@ public static class ArchiveReaderExtensions
         }
 
         transaction.Extract(entries.Keys, index => entries[index].Item2);
-        foreach (var (entry, stream) in entries.Values)
+        foreach (var (_, stream) in entries.Values)
         {
             stream.Seek(0, SeekOrigin.Begin);
         }
@@ -82,31 +111,5 @@ public static class ArchiveReaderExtensions
 
         entries.Clear();
         pool.Discard();
-    }
-
-    private static int GetStreamPoolCapacity(this ArchiveReader reader, int capacity)
-    {
-        if (capacity > 0)
-        {
-            return Math.Min(capacity, MaxCapacity);
-        }
-
-        return reader.GetUncompressedArchiveSize() switch
-        {
-            > MaxCapacity => MaxCapacity,
-            var value => (int)value
-        };
-    }
-
-    private static ulong GetUncompressedArchiveSize(this ArchiveReader reader)
-    {
-        return reader.Entries.Aggregate(0UL, (current, entry) => current + entry.UncompressedSize);
-    }
-
-    private static IScopedStreamPool CreateScopedStreamPool(bool useNativeMemory, int capacity)
-    {
-        return useNativeMemory
-            ? new Detail.NativeStreamPool(capacity)
-            : new Detail.DefaultStreamPool(capacity);
     }
 }
